okktrainer/ai/requirements.txt
python-dotenv==1.0.1
livekit-agents==1.4.2
livekit-plugins-silero==1.4.2

# локальные STT/TTS плагины:
git+https://github.com/CoreWorxLab/local-livekit-plugins.git
